requirements.txt
asgiref==3.7.2
Django==4.2.4
django-picklefield==3.1
django-q2==1.5.4
psutil==5.9.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
